% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_features.R
\name{calc_p_value}
\alias{calc_p_value}
\title{Get p value}
\usage{
calc_p_value(
  observed,
  expected,
  r = NULL,
  funcs = c("K", "G", "F"),
  obs_name = c(NA, "cor_name"),
  expect_name = c("rrl_", "cor_name"),
  K_cor = "trans",
  G_cor = "km",
  F_cor = "km",
  GXGH_cor = "km",
  GXHG_cor = "km",
  G2_cor = "km"
)
}
\arguments{
\item{observed}{a vector}

\item{expected}{a matrix or dataframe}
}
\description{
Get p value
}
