% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_features.R
\name{calc_T_val_envs}
\alias{calc_T_val_envs}
\title{T Value for envelope}
\usage{
calc_T_val_envs(
  relabelings,
  func = "K",
  rmin = 0,
  rmax = NA,
  sqrt = FALSE,
  rrl_name = "rrl_",
  K_cor = "trans",
  G_cor = "km",
  F_cor = "km",
  GXGH_cor = "km",
  GXHG_cor = "km"
)
}
\description{
T Value for envelope
}
