% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_summarize.R
\name{calc_T_vals}
\alias{calc_T_vals}
\title{Calculate T value}
\usage{
calc_T_vals(
  relabelings,
  func = "K",
  rmin = 0,
  rmax,
  K_cor = "border",
  G_cor = "km"
)
}
\arguments{
\item{relabelings}{object from \code{\link[rTEM]{relabel_summarize}}}

\item{func}{Which function is used: currently on G and K are supported}

\item{rmin}{what value to start the calculation from}

\item{rmax}{max value to run calculation to}

\item{K_cor}{boundary correction type for K}

\item{G_cor}{boundary correction type for G}
}
\description{
Takes input of relabeling object and calculates the T value for either K or G function
}
\details{
This uses the method from \emph{Baddeley et al.} to calculate the T value for an envelope for either
K  (\emph{K3est}) G \emph{G3est} function
}
\references{
Baddeley, A., Diggle, P. J., Hardegen, A., Lawrence, T_, Milne, R. K., & Nair, G. (2014).
On tests of spatial pattern based on simulation envelopes. Ecological Monographs, 84(3), 477–489. https://doi.org/10.1890/13-2042.1
}
