% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_analyze_TEM_feats.R
\name{combine_rank}
\alias{combine_rank}
\title{Combine ranks}
\usage{
combine_rank(tall_data, feature_name, funcs = c("G", "K"), im = 1)
}
\description{
Takes a tall formatted matrix and combines the ranks of feature `feat_name` for functions `funcs`
}
