% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_analyze_TEM_feats.R
\name{feat_list_to_wide}
\alias{feat_list_to_wide}
\title{Convert list of features to wide format}
\usage{
feat_list_to_wide(feats, names, col_names, feat_names)
}
\arguments{
\item{feats}{vector of feats}
}
\description{
Takes a list of lists and converts to single wide matrix
}
