% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_multimersim_inhibit.R
\name{findNearestPoints}
\alias{findNearestPoints}
\title{Find Points}
\usage{
findNearestPoints(OVER, UNDER, n = 1)
}
\arguments{
\item{OVER}{overlying point pattern}
}
\description{
Find Points
}
