% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_deprecated.R
\name{findNearestPoints_dep}
\alias{findNearestPoints_dep}
\title{Find Points}
\usage{
findNearestPoints_dep(OVER, UNDER, n = 1)
}
\arguments{
\item{OVER}{overlying point pattern}
}
\description{
Find Points
}
