% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gest_NN.R
\name{Gest_nn}
\alias{Gest_nn}
\title{Advanced Nearest Neighbour Distance Function G}
\usage{
Gest_nn(
  X,
  r = NULL,
  breaks = NULL,
  k = 1,
  ...,
  correction = c("rs", "km", "han"),
  domain = NULL
)
}
\arguments{
\item{k}{Integer, or integer vector. The algorithm will compute the distance to the kth nearest neighbour.
See \link[spatstat.geom]{nndist}}
}
\description{
Estimates the nearest neighbour distance distribution
Gest from a point pattern in a window of arbitrary shape. Modified from the \emph{spatstat} package
function \link[spatstat.core]{Gest}  to allow for usage of nearest neighbors beyond the first
}
