% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_deprecated.R
\name{get_features_deprecated}
\alias{get_features_deprecated}
\title{Get Features deprecated Dec 19 2024}
\usage{
get_features_deprecated(
  expected,
  simulated,
  funcs = c("G", "K", "F"),
  features = list(c("max_diff", "min_diff", "total_norm_diff", "total_norm_diff_squared",
    "total_diff", "total_squared_diff", "net_diff", "max_diff_envelope",
    "min_diff_envelope", "total_diff_envelope", "net_diff_envelope", "T_final",
    "T_final_ratio")),
  simulated_outer_name = "rrl_",
  simulated_inner_name = c("r", "mmean", "lo", "hi"),
  expected_inner_name = c(),
  sqrt = "K",
  K_cor = "trans",
  G_cor = "km",
  G2_cor = "km",
  G3_cor = "km",
  G4_cor = "km",
  F_cor = "km",
  GXGH_cor = "km",
  GXHG_cor = "km"
)
}
\arguments{
\item{expected}{a list of summary functions for the observed value}

\item{simulated}{a list of}

\item{funcs}{list of summary functions to get features for.  currently `G`, `K`, and `F` are supported}

\item{features}{a list.  Must be either length 1 or same length as `funcs` parameter
@description deprecated because the normalize total difference needed to only sum
non zero env width}
}
\description{
Get Features deprecated Dec 19 2024
}
