% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_hypothesis_testing.R
\name{get_global_env_test_expected}
\alias{get_global_env_test_expected}
\title{Global envelope test for list of list of curves}
\usage{
get_global_env_test_expected(expected, funcs, type, r_vals, sim_inds)
}
\arguments{
\item{expected}{list of matrices expected curve values.  Format: `expected[[func]]`,
func is the summary function name.}

\item{funcs}{vector containing summary function names.  These must also be the names of the inner list elements}

\item{type}{which type of `global_envelope_test` to perform}

\item{r_vals}{vector of length n values corresponding to radius (or other value) for each obs_mat and sim_mat curves}

\item{observed}{list of matrices of observed curve values.  Same format as expected curves}
}
\description{
Global envelope test for list of list of curves
}
