% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_new_funcs.R
\name{get_maxes}
\alias{get_maxes}
\title{Get max or min}
\usage{
get_maxes(
  envelope,
  observed,
  inds = NA,
  base_ind = 1,
  take = "max",
  funcs = c("G", "K", "F", "G2"),
  G_cor = "km",
  K_cor = "trans",
  F_cor = "km",
  G2_cor = "km",
  G3_cor = "km",
  G4_cor = "km",
  sqrt = "K"
)
}
\description{
Get max or min
}
