% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_analyze_TEM_feats.R
\name{get_png_square}
\alias{get_png_square}
\title{assemble png files into a png square}
\usage{
get_png_square(
  input_path,
  save_path,
  pat,
  to_name,
  y_space = 50,
  x_space = 50,
  resolution = 70
)
}
\description{
assemble png files into a png square
}
