% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_deprecated.R
\name{get_png_square_deprecated}
\alias{get_png_square_deprecated}
\title{assemble png files into a png square deprecated Dec 20 2024}
\usage{
get_png_square_deprecated(
  input_path,
  save_path,
  pat,
  to_name,
  y_space = 50,
  x_space = 50,
  resolution = 70
)
}
\description{
assemble png files into a png square deprecated Dec 20 2024
}
