% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_hypothesis_testing.R
\name{get_power_val}
\alias{get_power_val}
\title{Gett power value for two lists of lists of curves}
\usage{
get_power_val(
  expected,
  observed,
  thresh_val = 0.95,
  funcs = c("G", "K"),
  r_vals,
  n_relabs
)
}
\arguments{
\item{expected}{list of curves}

\item{observed}{list of curves}

\item{thresh_val}{threshold, or `1-significance level` of test}

\item{funcs}{vector of summary functions}

\item{n_relabs}{number of relabelings}

\item{r_values}{radius values corresponding to each observed and expected matrix}
}
\description{
Gett power value for two lists of lists of curves
}
