% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/under_construction.R
\name{get_ranks_dev}
\alias{get_ranks_dev}
\title{Get Ranks}
\usage{
get_ranks_dev(distance, weights)
}
\arguments{
\item{distance}{vector of distances}

\item{weights}{vector of weights.  Should be same length as distances}
}
\description{
For an input distances \eqn{d_1, d_2, d_3} and weights \eqn{w_1, w_2, w_3},
 returns weighted distance \eqn{\sqrt{w_1 d_1^2 + w_2 d_2^2 + w_3 d_3^2}}
 @export
}
