% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_analyze_TEM_feats.R
\name{get_tables}
\alias{get_tables}
\title{Get and save ordered tables for feature values of functions}
\usage{
get_tables(
  all_tall,
  func_groups,
  size_fracs,
  feature_name,
  name_image,
  image_num,
  path
)
}
\description{
Get and save ordered tables for feature values of functions
}
