% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_analyze_TEM_feats.R
\name{get_tall}
\alias{get_tall}
\title{Get tall data from list of lists}
\usage{
get_tall(
  iso_feats,
  vert_feats,
  feat_names,
  tall_names = names,
  all_funcs = all_funcs
)
}
\description{
Get tall data from list of lists
}
