% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_hypothesis_testing.R
\name{integrated_sq_dev}
\alias{integrated_sq_dev}
\title{Calculate squared area between two curves}
\usage{
integrated_sq_dev(x, mu, r)
}
\arguments{
\item{x}{observed curve}

\item{mu}{expected curve}

\item{r}{r values for both curves}
}
\description{
Calculate squared area between two curves
}
