% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_hypothesis_testing.R
\name{make_curve_set}
\alias{make_curve_set}
\title{Make curve set for observed and simulated data}
\usage{
make_curve_set(obs_mat, sim_mat, r_vals, ind)
}
\arguments{
\item{obs_mat}{n x m matrix of m curves each of length n}

\item{sim_mat}{n x j matrix of j curves each of length n}

\item{r_vals}{vector of length n values corresponding to radius (or other value) for each obs_mat and sim_mat curves}

\item{ind}{index (column number) of which obs_mat curve will be set as observed}
}
\description{
Make curve set for observed and simulated data
}
\details{
This function takes an observed and a simulated matrix and uses the curve_set function to turn it into a curve object to then be used in the global_envelope_test function
}
