% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_hypothesis_testing.R
\name{make_curve_set_expected}
\alias{make_curve_set_expected}
\title{Make curve set for observed data only}
\usage{
make_curve_set_expected(sim_mat, r_vals, obs_ind, sim_inds = 1:ncol(sim_mat))
}
\arguments{
\item{r_vals}{vector of length n values corresponding to radius (or other value) for each obs_mat curve}

\item{obs_mat}{n x m matrix of m curves each of length n}

\item{ind}{index (column number) of which obs_mat curve will be set as observed}
}
\description{
Make curve set for observed data only
}
\details{
This function takes an observed matrix and uses the curve_set function to turn it into a
curve object to then be used in the global_envelope_test function
}
