% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_multimer_simulation.R,
%   R/under_construction.R
\name{pick_neighbor}
\alias{pick_neighbor}
\title{Pick Neighbor}
\usage{
pick_neighbor(
  ranks,
  probs,
  distances = NA,
  sample_method = "rank",
  exponent = 1,
  group_size = 2
)

pick_neighbor(
  ranks,
  probs,
  distances = NA,
  sample_method = "rank",
  exponent = 1,
  group_size = 2
)
}
\arguments{
\item{ranks}{ranks for each point.  If based upon a weighted distance, then may not directly
correspond to the `distances`.}

\item{probs}{vector of probabilities.
For \eqn{probs = c(p_1, p_2, p_3, p_4)}, the probability of the first NN being
selected in \eqn{p_1}, the probability of the second is \eqn{p_2}, and so on}

\item{distances}{distance to each point.  Only used if \emph{sample_method = "exp"}
#' @param sample_method if equal to \emph{"rank"}, the probability of a point of rank \emph{x}
 being chosen as a guest is \emph{probs[x]}.  If equal to  \emph{"exp"},
 the probability of a point of rank \emph{x} being chosen
  as a guest is \emph{probs[x] * exp(-exponent * distances[ranks]))}}

\item{exponent}{a numeric. If \emph{sample_method = "exp"}, then
this is the value of \emph{exponent} in \emph{exp(-exponent * distances[ranks])}
#' @param group_size a numeric. How large will the groups be.  Pick `group_size -1` neighbors}
}
\description{
Pick Neighbor

Pick Neighbor
}
