% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_deprecated.R
\name{plot_summary_v01}
\alias{plot_summary_v01}
\title{Plot Summary functions}
\usage{
plot_summary_v01(
  func = "K",
  observed_values,
  envelopes,
  ...,
  pattern.colors = c(`95.0\% AI` = "pink", Median = "black", Observed = "blue"),
  fill.colors = pattern.colors,
  unit = "nm",
  K_cor = "trans",
  G_cor = "km",
  F_cor = "km",
  GXGH_cor = "km",
  GXHG_cor = "km"
)
}
\description{
Plot Summary functions
}
