% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_analyze_TEM_feats.R
\name{stack_png_to_png}
\alias{stack_png_to_png}
\title{Stack png files into a png file}
\usage{
stack_png_to_png(input_path, save_path, to_name, y_space = 50, resolution = 70)
}
\description{
Stack png files into a png file
}
