% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_features.R
\name{T_test}
\alias{T_test}
\title{Perform T Test}
\usage{
T_test(
  observed,
  expected,
  func = "G",
  envelope_value = NA,
  k = NA,
  m = NA,
  two_sided = TRUE,
  rmax = NA,
  rmin = 0,
  sqrt = FALSE,
  rrl_name = "rrl_",
  K_cor = "trans",
  G_cor = "km",
  F_cor = "km",
  GXGH_cor = "km",
  GXHG_cor = "km"
)
}
\arguments{
\item{k}{using kth largest and smallest values}

\item{m}{number of simulations/relabelings}
}
\description{
Perform T Test
}
