% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_new_funcs.R
\name{UPP_preprocess}
\alias{UPP_preprocess}
\title{Preprocess UPP data}
\usage{
UPP_preprocess(
  upp,
  guest_pattern,
  min_thick = NA,
  max_thick = NA,
  intensity_upp = 1,
  n_guests = NA
)
}
\description{
Preprocess UPP data
}
