% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_new_funcs.R
\name{UPP_preprocess_dev}
\alias{UPP_preprocess_dev}
\title{Preprocess UPP data (developement version)}
\usage{
UPP_preprocess_dev(
  upp,
  guest_pattern,
  min_thick = NA,
  max_thick = NA,
  intensity_upp = 1,
  n_guests = NA
)

UPP_preprocess_dev(
  upp,
  guest_pattern,
  min_thick = NA,
  max_thick = NA,
  intensity_upp = 1,
  n_guests = NA
)
}
\description{
Preprocess UPP data (developement version)

Preprocess UPP data
}
