% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_analyze_TEM_feats.R
\name{get_png_square}
\alias{get_png_square}
\title{assemble png files into a png square}
\usage{
get_png_square(
  input_path,
  save_path,
  pat,
  observed_pat = NA,
  row_file_names,
  to_name,
  y_space = 50,
  x_space = 50,
  resolution = 70,
  scale_image_by = 1,
  ind_order = c(5, 2, 3, 1),
  f_ind = 1,
  g_ind = 2,
  g2_ind = 3,
  g3_ind = 4,
  k_ind = 5
)
}
\description{
assemble png files into a png square
}
